api/requirements.txt
flask
flask-sqlalchemy
pyodbc
python-dotenv
